requirements.txt
torch
torchvision
tqdm
numpy
matplotlib
pyyaml